优化代码

+++ smart-http-server/src/main/java/org/smartboot/http/server/HttpServerConfiguration.java
@@ -322,7 +322,7 @@ public class HttpServerConfiguration {
         return wsIdleTimeout;
     }
 
-    public HttpServerConfiguration setWsIdleTimeout(int wsIdleTimeout) {
+    public HttpServerConfiguration setWsIdleTimeout(long wsIdleTimeout) {
         this.wsIdleTimeout = wsIdleTimeout;
         return this;
     }
